[Lint] Splits token into its own file like expr and statement
