+++ include/occa/lang/token/charToken.hpp
@@ -1,0 +1,30 @@
+#ifndef OCCA_LANG_TOKEN_CHARTOKEN_HEADER
+#define OCCA_LANG_TOKEN_CHARTOKEN_HEADER
+
+#include <occa/lang/token/token.hpp>
+
+namespace occa {
+  namespace lang {
+    class charToken : public token_t {
+    public:
+      int encoding;
+      std::string value;
+      std::string udf;
+
+      charToken(const fileOrigin &origin_,
+                int encoding_,
+                const std::string &value_,
+                const std::string &udf_);
+
+      virtual ~charToken();
+
+      virtual int type() const;
+
+      virtual token_t* clone() const;
+
+      virtual void print(io::output &out) const;
+    };
+  }
+}
+
+#endif
