+++ include/occa/lang/token/newlineToken.hpp
@@ -1,0 +1,23 @@
+#ifndef OCCA_LANG_TOKEN_NEWLINETOKEN_HEADER
+#define OCCA_LANG_TOKEN_NEWLINETOKEN_HEADER
+
+#include <occa/lang/token/token.hpp>
+
+namespace occa {
+  namespace lang {
+    class newlineToken : public token_t {
+    public:
+      newlineToken(const fileOrigin &origin_);
+
+      virtual ~newlineToken();
+
+      virtual int type() const;
+
+      virtual token_t* clone() const;
+
+      virtual void print(io::output &out) const;
+    };
+  }
+}
+
+#endif
