+++ include/occa/lang/token/primitiveToken.hpp
@@ -1,0 +1,28 @@
+#ifndef OCCA_LANG_TOKEN_PRIMITIVETOKEN_HEADER
+#define OCCA_LANG_TOKEN_PRIMITIVETOKEN_HEADER
+
+#include <occa/lang/token/token.hpp>
+
+namespace occa {
+  namespace lang {
+    class primitiveToken : public token_t {
+    public:
+      primitive value;
+      std::string strValue;
+
+      primitiveToken(const fileOrigin &origin_,
+                     const primitive &value_,
+                     const std::string &strValue_);
+
+      virtual ~primitiveToken();
+
+      virtual int type() const;
+
+      virtual token_t* clone() const;
+
+      virtual void print(io::output &out) const;
+    };
+  }
+}
+
+#endif
