+++ src/lang/token/variableToken.cpp
@@ -1,0 +1,25 @@
+#include <occa/lang/token/variableToken.hpp>
+#include <occa/lang/variable.hpp>
+
+namespace occa {
+  namespace lang {
+    variableToken::variableToken(const fileOrigin &origin_,
+                                 variable_t &variable) :
+      token_t(origin_),
+      value(variable) {}
+
+    variableToken::~variableToken() {}
+
+    int variableToken::type() const {
+      return tokenType::variable;
+    }
+
+    token_t* variableToken::clone() const {
+      return new variableToken(origin, value);
+    }
+
+    void variableToken::print(io::output &out) const {
+      out << value.name();
+    }
+  }
+}
